Run all batches in a loop as long as there are successes
We ding after every batch of 100 but keep going so long as the list keeps being non-empty and the list and delete endpoints keep returning the expected status codes.

+++ index.js
@@ -3,17 +3,30 @@ import owner from './owner.js';
 import pat from './pat.js';
 import repo from './repo.js';
 
-const url = `https://api.github.com/repos/${owner}/${repo}/actions/artifacts?per_page=100`;
-const response = await fetch(url, { headers: { Authorization: `Bearer ${pat}` } });
-const data = await response.json();
+let total;
+while (total === undefined || total > 0) {
+  const url = `https://api.github.com/repos/${owner}/${repo}/actions/artifacts?per_page=100`;
+  const response = await fetch(url, { headers: { Authorization: `Bearer ${pat}` } });
+  if (response.status !== 200) {
+    child_process.exec('afplay /System/Library/Sounds/Sosumi.aiff');
+    process.exit(1);
+  }
 
-let counter = 0;
-for (const artifact of data.artifacts) {
-  const url = `https://api.github.com/repos/${owner}/${repo}/actions/artifacts/${artifact.id}`;
-  const response = await fetch(url, { method: 'DELETE', headers: { Authorization: `Bearer ${pat}` } });
-  const rate = `${response.headers.get('X-RateLimit-Remaining')}/${response.headers.get('X-RateLimit-Limit')}`
-  console.log(`${counter++}/${data.total_count}: ${artifact.id} - ${artifact.name}: ${response.status} ${response.statusText} (${rate})`);
+  const data = await response.json();
+  total = data.total_count;
+
+  let counter = 0;
+  for (const artifact of data.artifacts) {
+    const url = `https://api.github.com/repos/${owner}/${repo}/actions/artifacts/${artifact.id}`;
+    const response = await fetch(url, { method: 'DELETE', headers: { Authorization: `Bearer ${pat}` } });
+    if (response.status !== 204) {
+      child_process.exec('afplay /System/Library/Sounds/Sosumi.aiff');
+    }
+
+    const rate = `${response.headers.get('X-RateLimit-Remaining')}/${response.headers.get('X-RateLimit-Limit')}`
+    console.log(`${counter++}/${total}: ${artifact.id} - ${artifact.name}: ${response.status} ${response.statusText} (${rate})`);
+  }
+
+  // Beep to signal batch completion (macOS only)
+  child_process.exec('afplay /System/Library/Sounds/Glass.aiff');
 }
-
-// Beep to signal completion (macOS only)
-child_process.exec('afplay /System/Library/Sounds/Glass.aiff');
